Add case and stencil puck stls
Add new case stl.
Add the PCB stencil jig 'puck'. This is a jig with holes for 2.54mm pins that line up with the programming pin holes in the PCB to make lining up with the solder paste screen easier. This is 18mm thick (standard MDF thickness) and 35mm in diameter (standard spade drill bit diamater).